app/requirements.txt
garth
jpype1
zwift-client
garminconnect